app/python/src/requirements.txt
flask
werkzeug
numpy
Pillow